Add FreeMote toolkit, psb build/decompilation and compile script_body

+++ make.py
@@ -24,13 +24,22 @@ parser_script.add_argument("-d", action="store_true", help="decompile scripts")
 parser_script.add_argument("-r", type=str, help="folder for redecompiled scripts (-cd for roundtrip)")
 
 
+parser_man = subparsers.add_parser("man", help="data manipulation")
+man_subparsers = parser_man.add_subparsers()
+
+parser_psb = man_subparsers.add_parser("psb", help="manipulate psb")
+parser_psb.add_argument("input", type=str, help="input json or binary")
+parser_psb.add_argument("output", type=str, help="output folder or binary")
+parser_psb.add_argument("-c", action="store_true", help="compile (build)")
+parser_psb.add_argument("-d", action="store_true", help="decompile (build)")
+
 # config
 PSB_KEY = "38757621acf82"
 
 NUTCRACKER_PATH = "tools/nutcracker.exe"
 SQ_PATH = "tools/sq.exe"
 PSB_DECOMPILE_PATH = "tools/FreeMoteToolkit/PsbDecompile.exe"
-PSB_COMPILE_PATH = "tools/FreeMoteToolkit/PsbCompile.exe"
+PSB_BUILD_PATH = "tools/FreeMoteToolkit/PsBuild.exe"
 
 BUILD_ROOT = "build"
 TMP_ROOT = f"{BUILD_ROOT}/tmp"
@@ -39,6 +48,8 @@ FILES_ROOT = f"{BUILD_ROOT}/windata"
 CACHE_FILE = f"{BUILD_ROOT}/cache.json"
 
 SCRIPTS_SRC = f"script/src"
+SCRIPT_JSON = f"script/psb-json"
+
 
 hash_store: HashStorage
 
@@ -63,6 +74,40 @@ def check_dir(directory):
     return changed
 
 
+def build_psb(src, dst):
+    print(f"building {src}")
+    result = run_single(PSB_BUILD_PATH, "info-psb", "-k", PSB_KEY, "-p", "win", src)
+    if result:
+        psb_name = os.path.split(src)[1].split("_")[0]
+        files_to_move = [
+            f"{psb_name}_info.psb.m",
+            f"{psb_name}_body.bin"
+        ]
+        for file in files_to_move:
+            shutil.move(file, f"{dst}/{file}")
+        print(f"{src} built succesfully")
+    else:
+        print(f"failed to build {src}")
+
+
+def decompile_psb(src, dst):
+    print(f"decompiling {src}")
+    result = run_single(PSB_DECOMPILE_PATH, "info-psb", "-k", PSB_KEY, src)
+    if result:
+        root_dir = os.path.split(src)[0]
+        psb_name = os.path.split(src)[1].split("_")[0]
+        files_to_move = [
+            psb_name,
+            f"{psb_name}_info.psb.m.json",
+            f"{psb_name}_info.psb.m.resx.json"
+        ]
+        for file in files_to_move:
+            shutil.move(f"{root_dir}/{file}", f"{dst}/{file}")
+        print(f"{src} decompiled succesfully")
+    else:
+        print(f"failed to decompile {src}")
+
+
 def compile_script(src, dst):
     print(f"compiling {src}")
     result = run_single(SQ_PATH, "-c", "-o", dst, src)
@@ -84,32 +129,40 @@ def decompile_script(src, dst):
 def compile_scripts():
     print("compiling scripts...")
     tmp_path = f"{TMP_ROOT}/scripts"
-    ensure_path(tmp_path)
+    ensure_path(f"{tmp_path}/script")
+    changed = False
     for script_file in os.listdir(SCRIPTS_SRC):
         src_path = f"{SCRIPTS_SRC}/{script_file}"
-        out_path = f"{tmp_path}/{script_file}.m"
+        out_path = f"{tmp_path}/script/{script_file}.m"
         if hash_store.check_changed(src_path) or not os.path.exists(out_path):
             compile_script(src_path, out_path)
             hash_store.update_file(src_path)
+            changed = True
     hash_store.save()
+    if changed:
+        ensure_path(FILES_ROOT)
+        info_json = "script_info.psb.m.json"
+        shutil.copy(f"{SCRIPT_JSON}/{info_json}", tmp_path)
+        shutil.copy(f"{SCRIPT_JSON}/script_info.psb.m.resx.json", tmp_path)
+        build_psb(f"{tmp_path}/{info_json}", f"{FILES_ROOT}")
 
 
 def make_main(args):
     global hash_store
 
-    ensure_path(BUILD_ROOT)
-    hash_store = HashStorage(CACHE_FILE)
-
     if args.c:
         clean()
+
+    ensure_path(BUILD_ROOT)
+    hash_store = HashStorage(CACHE_FILE)
 
     if args.x:
         compile_scripts()
 
-    if not args.t:
-        shutil.rmtree(TMP_ROOT)
-
     hash_store.save()
+
+    if not args.t and os.path.exists(TMP_ROOT):
+        shutil.rmtree(TMP_ROOT)
 
 
 def script_main(args):
@@ -137,9 +190,27 @@ def script_main(args):
             decompile_script(src_path, dst_path)
 
 
+def psb_main(args):
+
+    if args.c and args.d:
+        print("choose one action")
+        return
+
+    ensure_path(args.output)
+
+    if args.d:
+        decompile_psb(args.input, args.output)
+        return
+
+    if args.c:
+        build_psb(args.input, args.output)
+        return
+
+
 if __name__ == "__main__":
     parser_make.set_defaults(func=make_main)
     parser_script.set_defaults(func=script_main)
+    parser_psb.set_defaults(func=psb_main)
     argss = parser.parse_args()
     if "func" not in argss:
         parser.print_help()
